Removed the dependency on Hive's old SerDeUtils.lookupDeserialzier()
Since Hive >= 0.13 doesn't use this method, and this library really only needed it to load serde classes, I was able to replace the dependency by putting the operative code directly in the HCatScheme#createSerDe() method. The main difference is that this change removes the use of Hive's internal serde cache, but it wasn't being populated with anything that couldn't just be loaded directly anyway. That implementation detail could be added back here if it were needed.